feat: fmt/clippy CIジョブ追加 + 全体整形・clippy解消 + 負パステスト拡充

+++ crates/elab/src/width.rs
@@ -4,7 +4,7 @@ use crate::ElabError;
 
 pub fn infer_width(expr: &Expr, context_width: Option<u32>) -> Result<u32, ElabError> {
     match expr {
-        Expr::Const(val) => Ok(val.width() as u32),
+        Expr::Const(val) => Ok(val.width()),
         _ => Ok(context_width.unwrap_or(1)),
     }
 }
